Another 'size' fix - 'long' to 'int'
I guess I missed this from the earlier set of 'fixes'.  In DCW
there's only byte, short, and 32-bit integers ... no such thing
as 64 bit back then.  So the mystery remains, why was I using
'unisgned long' and 'long'.  It worked because all the compilers
I was using implemented them as 32 bit.  Now, in 2026, a 'long'
really is 64-bit.  And that caused parse errors.

So this one ('long') I missed.

Changes to be committed:
	modified:   include/gmsFile/gmsIndexFile.h

+++ C_source/include/gmsFile/gmsIndexFile.h
@@ -52,7 +52,7 @@
 
    typedef struct _indexRecordType
    {
-       long byteOffset;
+       int  byteOffset;
        int  numBytes;
    } indexRecordType;
